+++ R/tmb-sim.R
@@ -332,8 +332,12 @@ sdmTMB_simulate <- function(formula,
 #'   effects (this only simulates observation error). `~0` or `NA` to simulate
 #'   new random affects (smoothers, which internally are random effects, will
 #'   not be simulated as new).
+#' @param mle_mvn_samples Applies if `type = "mle-mvn"`. If `"single"`, take
+#'   a single MVN draw from the random effects. If `"multiple"`, take an MVN
+#'   draw from the random effects for each of the `nsim`.
 #' @param model If a delta/hurdle model, which model to simulate from?
 #'   `NA` = combined, `1` = first model, `2` = second mdoel.
+#' @param newdata Optional new data frame from which to simulate.
 #' @param mcmc_samples An optional matrix of MCMC samples. See `extract_mcmc()`
 #'   in the \href{https://github.com/pbs-assess/sdmTMBextra}{sdmTMBextra}
 #'   package.
@@ -381,10 +385,16 @@ sdmTMB_simulate <- function(formula,
 simulate.sdmTMB <- function(object, nsim = 1L, seed = sample.int(1e6, 1L),
                             type = c("mle-eb", "mle-mvn"),
                             model = c(NA, 1, 2),
-                            re_form = NULL, mcmc_samples = NULL, silent = FALSE, ...) {
+                            newdata = NULL,
+                            re_form = NULL,
+                            mle_mvn_samples = c("single", "multiple"),
+                            mcmc_samples = NULL,
+                            silent = FALSE,
+                            ...) {
   set.seed(seed)
   type <- tolower(type)
   type <- match.arg(type)
+  mle_mvn_samples <- match.arg(mle_mvn_samples)
   assert_that(as.integer(model[[1]]) %in% c(NA_integer_, 1L, 2L))
 
   # need to re-attach environment if in fresh session
@@ -403,6 +413,16 @@ simulate.sdmTMB <- function(object, nsim = 1L, seed = sample.int(1e6, 1L),
     stopifnot(length(object$tmb_data$sim_re) == 6L) # in case this gets changed
     tmb_dat$sim_re <- c(rep(1L, 5L), 0L) # last is smoothers; don't simulate them
   }
+
+  if (!is.null(newdata)) {
+    # generate prediction TMB data list
+    p <- predict(object, newdata = newdata, return_tmb_data = TRUE, ...)
+    # move data elements over
+    p <- move_proj_to_tmbdat(p, object, newdata)
+    p$sim_re <- tmb_dat$sim_re
+    tmb_dat <- p
+  }
+
   newobj <- TMB::MakeADFun(
     data = tmb_dat, map = object$tmb_map,
     random = object$tmb_random, parameters = object$tmb_obj$env$parList(), DLL = "sdmTMB"
@@ -411,9 +431,17 @@ simulate.sdmTMB <- function(object, nsim = 1L, seed = sample.int(1e6, 1L),
   # params MLE/MVN stuff
   if (is.null(mcmc_samples)) {
     if (type == "mle-mvn") {
-      new_par <- .one_sample_posterior(object)
+      if (mle_mvn_samples == "single") {
+        new_par <- .one_sample_posterior(object)
+        new_par <- replicate(nsim, new_par)
+      } else {
+        new_par <- lapply(seq_len(nsim), \(i) .one_sample_posterior(object))
+        new_par <- do.call(cbind, new_par)
+      }
     } else if (type == "mle-eb") {
       new_par <- object$tmb_obj$env$last.par.best
+      new_par <- lapply(seq_len(nsim), \(i) new_par)
+      new_par <- do.call(cbind, new_par)
     } else {
       cli_abort("`type` type not defined")
     }
@@ -432,7 +460,7 @@ simulate.sdmTMB <- function(object, nsim = 1L, seed = sample.int(1e6, 1L),
   } else {
     for (i in seq_len(nsim)) {
       if (!silent) cli::cli_progress_update()
-      ret[[i]] <- newobj$simulate(par = new_par, complete = FALSE)$y_i
+      ret[[i]] <- newobj$simulate(par = new_par[, i, drop = TRUE], complete = FALSE)$y_i
     }
   }
   if (!silent) cli::cli_progress_done()