+++ tests/testthat/test-6-tmb-simulation.R
@@ -180,6 +180,51 @@ test_that("simulate() behaves OK with or without random effects across types", {
   expect_length(s, 969)
 })
 
+test_that("simulate() method works with newdata", {
+  skip_on_cran()
+  fit <- sdmTMB(
+    present ~ 1,
+    time = "year",
+    data = pcod_2011, spatial = "on",
+    spatiotemporal = "iid",
+    family = binomial(),
+    mesh = pcod_mesh_2011
+  )
+  s <- simulate(fit)
+  expect_true(nrow(s) == nrow(pcod_2011))
+  g <- replicate_df(qcs_grid, "year", unique(pcod_2011$year))
+  s <- simulate(fit, newdata = g)
+  expect_true(nrow(s) == nrow(g))
+  s <- simulate(fit, newdata = subset(g, year == 2011))
+  nrow(s)
+  expect_true(nrow(s) == nrow(subset(g, year == 2011)))
+
+  gg <- subset(g, year == 2011)
+  set.seed(1)
+  s <- simulate(fit, newdata = gg, nsim = 400L)
+  a <- apply(s, 1, mean)
+  p <- predict(fit, newdata = gg)
+  plot(a, plogis(p$est))
+  expect_gt(cor(a, plogis(p$est)), 0.98)
+
+  set.seed(1)
+  s1 <- simulate(fit, type = "mle-mvn", mle_mvn_samples = "single", nsim = 100)
+  set.seed(1)
+  s2 <- simulate(fit, type = "mle-mvn", mle_mvn_samples = "multiple", nsim = 100)
+  set.seed(1)
+  s3 <- simulate(fit, type = "mle-eb", nsim = 100)
+
+  expect_false(identical(s1, s2))
+  expect_false(identical(s1, s3))
+
+  sd1 <- apply(s1, 1, sd)
+  sd2 <- apply(s2, 1, sd)
+  sd3 <- apply(s3, 1, sd)
+
+  expect_lt(mean(sd1), mean(sd2))
+})
+
+
 # test_that("TMB Delta simulation works", {
 #   skip_on_cran()
 #   skip_if_not_installed("INLA")